Improve the Sphinx configs for laTeX/PDF builds (#12662)
* Improve the Sphinx configs for Latex/PDF builds

This also cleans up a bunch of RST and UML formats and files, which
should work and improve HTML output as well.

We also try to automate more of the builds, setting up the appropraite
pyenv and also making sure we deal with macOS app focus.

Finally, this upgrades PlantUML to v1.2025.10.

Build improvements

* Cleaning up some existing docs for formatting

* More TOC reorgs

* Chapter page breaks