ATSReplayTest: autests via replay.yaml (#12648)
For context, this continues the work started in #12371. The goal is to
take full advantage of the autest extension capabilities to make writing
ATS autests easier.

This adds the Test.ATSReplayTest autest extensions so that tests can be
written through a replay.yaml file that specifies both the configuration
of the test processes as well as the HTTP Proxy Verifier traffic for the
test.

This patch updates the header_rewrite_bundle to use this new framework.

(cherry picked from commit 4b3354fe383afbc2cd08d0dd1a87334f345735ec)