Add a per client connection max exempt list cript (#12476)
This updates our build system to be able to make pre-compiled cripts.
Thus, adding a cript is now as easy as adding a plugin. Simply use
add_cript instead of add_atsplugin in CMakeLists.txt, and it will build
your cript for you.

As a part of this, this adds connection_exempt_list.cript, our first
cript plugin that sets the per client exempt list.